modify NetworkAutofacModule
