+++ AElf.Concurrency.Worker/Program.cs
@@ -4,6 +4,7 @@ using AElf.Execution;
 using AElf.Kernel;
 using AElf.Kernel.Modules.AutofacModule;
 using AElf.Miner;
+using AElf.Network;
 using AElf.Runtime.CSharp;
 using AElf.SmartContract;
 using Autofac;
@@ -71,7 +72,7 @@ namespace AElf.Concurrency.Worker
             builder.RegisterModule(new TransactionManagerModule());
             builder.RegisterModule(new LoggerModule());
             builder.RegisterModule(new DatabaseAutofacModule());
-            builder.RegisterModule(new NetworkModule(isMiner));
+            builder.RegisterModule(new NetworkAutofacModule());
             builder.RegisterModule(new MinerAutofacModule(null));
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new StorageModule());
