+++ AElf.Launcher/Program.cs
@@ -105,7 +105,7 @@ namespace AElf.Launcher
             builder.RegisterModule(new TransactionManagerModule());
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new LoggerModule("aelf-node-" + NetworkConfig.Instance.ListeningPort));
-            builder.RegisterModule(new NetworkModule(isMiner));
+            builder.RegisterModule(new NetworkAutofacModule());
             builder.RegisterModule(new RpcServicesModule());
             builder.RegisterModule(new StorageModule());
             builder.RegisterType<ChainService>().As<IChainService>();
